api/requirements.txt
aiocryptopay==0.4.5
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
Babel==2.9.1
cached-property==1.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.112.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflection==0.5.1
magic-filter==1.0.12
multidict==6.0.5
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
pytonapi==0.3.9
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlalchemy-orm==1.2.10
starlette==0.38.4
StrEnum==0.4.15
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
websockets==13.0.1
yarl==1.9.7
PyJWT==2.9.0
